Add tests for closed programmes, merge legacy tests into one file
